修正了MySoft.Data中复杂关联查询的bug，重构了数据关联查询的部分业务逻辑！

+++ MySoftSolutionV3/MySoft.Data/Common/ILogable.cs
@@ -23,6 +23,6 @@ namespace MySoft.Data
         /// <param name="parameter"></param>
         /// <param name="result"></param>
         /// <param name="elapsedTime"></param>
-        void End(string cmdText, SQLParameter[] parameter, object result, int elapsedTime);
+        void End(string cmdText, SQLParameter[] parameter, object result, long elapsedTime);
     }
 }
